pyproject.toml
[project]
name = "sunmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.8.6",
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.1",
    "pgvector<0.4",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
